More migration towards using SuitOrdinal enum.

+++ Seahaven2e/CardLocation.h
@@ -119,7 +119,7 @@ public:
    bool operator!=(CardLocation location) const;
 
 public:
-   __attribute__((noinline)) static CardLocation AcePile(Suit suit) {  return CardLocation(Region::Ace, suit.GetOrdinal()); }
+   __attribute__((noinline)) static CardLocation AcePile(Suit suit) {  return CardLocation(Region::Ace, (uint8_t)suit.GetOrdinal()); }
    __attribute__((noinline)) static CardLocation Column(uint8_t column, uint8_t index) { return CardLocation((Region)column, index); }
    __attribute__((noinline)) static inline CardLocation Tower(uint8_t index) { return CardLocation(Region::Tower, index); }
    static constexpr CardLocation Null() { return CardLocation(Region::Null, 0); }
